tests: test the main action runner
