+++ .github/actions/semantic-pr-footer-v1/src/run.ts
@@ -1,0 +1,42 @@
+import type core from '@actions/core'
+import isValidFooter from './is-valid-footer'
+
+export type Core = Pick<typeof core, 'setFailed' | 'info'>
+export type PayLoad = {
+  pull_request: {
+    body: string
+  }
+}
+
+export default function run(
+  core: Core,
+  payload?: PayLoad
+) {
+  try {
+    const body: string | undefined =
+      payload &&
+      payload.pull_request &&
+      payload.pull_request.body
+
+    if (!body) {
+      core.setFailed('PR does not have a body')
+      return
+    }
+
+    const bodyLines = body.split(/[\r\n]+/)
+    const footer = bodyLines[bodyLines.length - 1]
+
+    core.info(`Validating PR footer: "${footer}"`)
+
+    if (!isValidFooter(footer)) {
+      core.setFailed(
+        'PR footer does not close an issue (`Closes: `), reference an issue (`Ref: ` or `Refs: `), provide QA notes (`QA notes: `), or state that no QA is needed (`No QA needed` or `No QA required`)'
+      )
+      return
+    }
+
+    core.info('Footer matches team policy')
+  } catch (error) {
+    core.setFailed((error as Error).message)
+  }
+}
