+++ package.json
@@ -23,6 +23,7 @@
     "@actions/github": "^5.1.1",
     "@types/chai": "^4.3.4",
     "@types/mocha": "^10.0.0",
+    "@types/sinon": "^10.0.16",
     "@typescript-eslint/eslint-plugin": "^5.43.0",
     "@typescript-eslint/parser": "^5.43.0",
     "@vercel/ncc": "^0.36.1",
@@ -33,6 +34,7 @@
     "mocha": "^10.1.0",
     "nyc": "^15.1.0",
     "prettier": "^2.7.1",
+    "sinon": "^15.2.0",
     "ts-node": "^10.9.1",
     "typescript": "4.8"
   },
